Add Slack connector, DOCX/XLSX support, and server-side credential flow
- Add Slack connector with bot token loaded from SLACK_BOT_TOKEN env var
- Add POST /connectors/{id}/connect endpoint for server-side credential flow
  (no token input from UI; extensible for future non-OAuth connectors)
- Add DOCX extraction via python-docx and XLSX via openpyxl in Google Drive fetcher
- Add ConnectorDetailPage with sync history, documents, and settings tabs
- Fix auto-sync: new connectors no longer sync automatically on connect
- Add clear sync history with confirmation modal
- Remove Documents/Last synced stats from connector tiles
- Hide idle status badge
- Add Slack connector icon and ConnectorIcon component
- Update frontend to server_env auth flow (no BotTokenModal)
- Add test documents for financial data

+++ backend/app/services/search_service.py
@@ -191,7 +191,7 @@ async def _apply_semantic_rerank(
         doc_id = str(row["document_id"])
         best = best_chunk_by_doc.get(doc_id)
         if best is None or sim > best["similarity"]:
-            best_chunk_by_doc[doc_id] = {"similarity": sim, "content": row["content"]}
+            best_chunk_by_doc[doc_id] = {"similarity": float(sim), "content": row["content"]}
 
     lexical_max = max((r["score"] for r in results), default=0.0)
     if lexical_max <= 0:
@@ -204,7 +204,7 @@ async def _apply_semantic_rerank(
         semantic_norm = max(0.0, min(1.0, (semantic["similarity"] + 1) / 2)) if semantic else 0.0
         reranked.append({
             **r,
-            "score": lexical_norm * 0.6 + semantic_norm * 0.4,
+            "score": float(lexical_norm * 0.6 + semantic_norm * 0.4),
             "snippet": semantic["content"][:320] if semantic else r["snippet"],
         })
